Suggested changes to avoid funnel breaks
In the case where a user has an active SPiD session, but has not accepted the product agreement, the SPiD SDK will never return the anonymous tracking uid used by Mixpanel. This is because the SDK never fires the "auth.visitor" event even though the data was received from SPiD. The tracking Id is returned if the user has no SPiD session at all, so it should also be fired in the case above.  

Beause of the problem above, Mixpanel creates its own tracking id. This causes funnel breaks on a big portion of the traffic as soon as the visit a SPiD page.

I have also added a SPiD status variable for convenience. It is really useful to know what kind of conversion funnel the user must follow to convert. Right now we do not have a way of segment conversion rates for no SPiD users, SPiD users that has not accepted the agreement, current users etc.